pad meshlet and change to 4bytes uint

+++ src/vk_mesh.h
@@ -23,8 +23,8 @@ struct vertex {
 };
 
 struct meshlet {
-    uint16_t vertex_index[64];
-    uint16_t indices[378];
+    uint32_t vertex_index[64];
+    uint32_t indices[378];
     uint32_t vertex_count;
     uint32_t index_count;
 };
